Fixes #12
--APIv1ExampleRunnerSendDeliveryNoticeToCustomer.cs
Added example class to show an example of how use the Send Delivery Notice To Customer API endpoint class to create a delivery notice and sent it to a customer in Squizz.

--APIv1EndpointOrgSendDeliveryNoticeToCustomer.cs
Added class to handle calling the SQUIZZ platform API endpoint for sending a delivery notice to a customer.

--APIv1Constants.cs
Added constant to store the relative URL of the Send Delivery Notice To Customer API endpoint.

--APIv1ExampleRunner.cs
Added ability to call the Send Delivery Notice To Customer example class in the example console application.

--AssemblyInfo.cs
--SquizzPlatformAPILibrary.nuspec
Updated library version